requirements.txt
django==4.2.5
django-debug-toolbar==4.3.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
Pillow==10.0.1
drf-spectacular==0.27.1
flake8==7.0.0
flake8-quotes==3.3.2
flake8-variables-names==0.0.6
pep8-naming==0.13.3
isort==5.13.2
